+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - L2CAP ERTM: avoid read-after-free on decline of incoming connection
 - GATT Client: set uuid16 to zero when deserializing uuid128 services, characteristics, and descriptors
- 
+- att_db_util: fix realloc of att db buffer for large attributes
+
 ### Added
 - att_db_util: provide GATT Database Hash via att_db_util_hash_calc
 - GATT Compiler: provide GATT Database Hash via GATT_DATABASE_HASH Characteristic